Reorganization for new drawCase.go
Reorganization for new drawCase.go
